src/tool/mcp_tools/browser_oi_tool/pyproject.toml
[tool.poetry]
name = "browser_oi_tool"
version = "0.1.0"
description = "MCP container exposing Open Interpreter via FastMCP as browser use tool"
authors = ["Dmytro Rohovyi"]

[tool.poetry.dependencies]
python = "3.12.10"

# MCP server
fastmcp = "^2.0.0"
open-interpreter = "0.4.3"
loguru = "^0.7.2"

# Browser automation
selenium = "^4.17.0"
playwright = {extras = ["chromium"], version = "^1.41.0"} # prefered
undetected-chromedriver = "^3.2.1" # with selenium for bot checks
requests = "^2.32.0"

# GUI automation
pyautogui = "^0.9.53"
pillow = "^10.1.0"

# Optional: image/screenshot comparison or data handling
opencv-python = "^4.8.1"
pandas = "^2.2.0"
websocket-client = "^1.6.0"

aiofiles = "^23.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
